Implemented Stripe Payment Gateway with refund functionality and final few ADMIN dashboard based APIs

+++ src/main/java/com/sameerahmed/projects/airBnbApp/service/HotelServiceImpl.java
@@ -2,20 +2,30 @@ package com.sameerahmed.projects.airBnbApp.service;
 
 import com.sameerahmed.projects.airBnbApp.dto.HotelDto;
 import com.sameerahmed.projects.airBnbApp.dto.HotelInfoDto;
+import com.sameerahmed.projects.airBnbApp.dto.HotelReportDto;
 import com.sameerahmed.projects.airBnbApp.dto.RoomDto;
 import com.sameerahmed.projects.airBnbApp.entity.Hotel;
 import com.sameerahmed.projects.airBnbApp.entity.Room;
+import com.sameerahmed.projects.airBnbApp.entity.User;
 import com.sameerahmed.projects.airBnbApp.exception.ResourceNotFoundException;
+import com.sameerahmed.projects.airBnbApp.exception.UnAuthorizedException;
 import com.sameerahmed.projects.airBnbApp.repository.HotelRepository;
 import com.sameerahmed.projects.airBnbApp.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.sameerahmed.projects.airBnbApp.util.AppUtils.getCurrentUser;
 
 @Service
 @Slf4j
@@ -32,6 +42,10 @@ public class HotelServiceImpl implements HotelService {
         log.info("Creating a new hotel with name: {}", hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        hotel.setOwner(user);
         Hotel savedHotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with ID: {}", hotelDto.getId());
         return modelMapper.map(savedHotel, HotelDto.class);
@@ -43,6 +57,11 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizedException("This user does not own this hotel with id: " + id);
+        }
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -51,6 +70,12 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository.findById(id)
                 .orElseThrow(() ->
                         new ResourceNotFoundException("Hotel not found with id: " + id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizedException("This user does not own this hotel with id: " + id);
+        }
         modelMapper.map(hotelDto, hotel);
         hotel.setAmenities(hotelDto.getAmenities());
         hotel.setPhotos(hotelDto.getPhotos());
@@ -65,6 +90,11 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizedException("This user does not own this hotel with id: " + id);
+        }
         for (Room room : hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
@@ -79,6 +109,11 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizedException("This user does not own this hotel with id: " + id);
+        }
         hotel.setActive(true);
 //        assuming only done once
         for (Room room : hotel.getRooms()) {
@@ -97,6 +132,18 @@ public class HotelServiceImpl implements HotelService {
                 .map((element) -> modelMapper.map(element, RoomDto.class))
                 .toList();
 
-        return new HotelInfoDto(modelMapper.map(hotel,HotelDto.class),rooms);
+        return new HotelInfoDto(modelMapper.map(hotel, HotelDto.class), rooms);
     }
+
+    @Override
+    public List<HotelDto> getAllHotels() {
+        User user = getCurrentUser();
+        log.info("Getting all hotels for the admin user with ID: {}", user.getId());
+
+        List<Hotel> hotels = hotelRepository.findByOwner(user);
+        return hotels.stream()
+                .map((element) -> modelMapper.map(element, HotelDto.class))
+                .collect(Collectors.toList());
+    }
+
 }
